chainctl-image-diff: use hardcoded identity ids in cosign verify
It used to be the case that calling `chainctl iam account-associations describe`
without an org name would work here, presuambly because the identity
we're using only has/had access to one org.

Now it seems to push you into the multi-select workflow which fails,
naturally, because we're not running in an interactive session.

We don't need to fetch these values on the fly, so fix this by hardcoding them.